Reorganize import from DataViews: import only from directory
