+++ packages/edit-site/src/components/dataviews/index.js
@@ -1,2 +1,9 @@
 export { default as DataViews } from './dataviews';
-export { VIEW_LAYOUTS } from './constants';
+export {
+	VIEW_LAYOUTS,
+	LAYOUT_GRID,
+	LAYOUT_TABLE,
+	LAYOUT_SIDE_BY_SIDE,
+	ENUMERATION_TYPE,
+	OPERATOR_IN,
+} from './constants';
